+++ app/src/main/java/org/breezyweather/sources/dmi/DmiService.kt
@@ -20,10 +20,15 @@ import android.content.Context
 import android.graphics.Color
 import io.reactivex.rxjava3.core.Observable
 import org.breezyweather.common.basic.models.Location
+import org.breezyweather.common.basic.wrappers.SecondaryWeatherWrapper
 import org.breezyweather.common.basic.wrappers.WeatherWrapper
+import org.breezyweather.common.exceptions.InvalidLocationException
+import org.breezyweather.common.exceptions.SecondaryWeatherException
 import org.breezyweather.common.source.HttpSource
 import org.breezyweather.common.source.MainWeatherSource
+import org.breezyweather.common.source.ParameterizedLocationSource
 import org.breezyweather.common.source.ReverseGeocodingSource
+import org.breezyweather.common.source.SecondaryWeatherSource
 import org.breezyweather.common.source.SecondaryWeatherSourceFeature
 import org.breezyweather.sources.dmi.json.DmiResult
 import org.breezyweather.sources.dmi.json.DmiWarningResult
@@ -32,7 +37,8 @@ import javax.inject.Inject
 
 class DmiService @Inject constructor(
     client: Retrofit.Builder
-) : HttpSource(), MainWeatherSource, ReverseGeocodingSource {
+) : HttpSource(), MainWeatherSource, SecondaryWeatherSource,
+    ReverseGeocodingSource, ParameterizedLocationSource {
 
     override val id = "dmi"
     override val name = "Danmarks Meteorologiske Institut (DMI)"
@@ -59,12 +65,20 @@ class DmiService @Inject constructor(
         val weather = mApi.getWeather(
             location.latitude,
             location.longitude,
-            "llj"
+            DMI_WEATHER_CMD
         )
 
         val alerts = if (!ignoreFeatures.contains(SecondaryWeatherSourceFeature.FEATURE_ALERT) &&
-            location.countryCode.equals("DK", ignoreCase = true) && !location.cityId.isNullOrEmpty()) {
-            mApi.getAlerts(location.cityId)
+            location.countryCode.equals("DK", ignoreCase = true)
+        ) {
+            val id = location.parameters.getOrElse(id) { null }?.getOrElse("id") { null }
+            if (!id.isNullOrEmpty()) {
+                mApi.getAlerts(id)
+            } else {
+                Observable.create { emitter ->
+                    emitter.onNext(DmiWarningResult())
+                }
+            }
         } else {
             Observable.create { emitter ->
                 emitter.onNext(DmiWarningResult())
@@ -81,24 +95,88 @@ class DmiService @Inject constructor(
         }
     }
 
+
+    // SECONDARY WEATHER SOURCE
+    override val supportedFeatures = listOf(
+        SecondaryWeatherSourceFeature.FEATURE_ALERT
+    )
+    override fun isFeatureSupportedForLocation(
+        feature: SecondaryWeatherSourceFeature, location: Location
+    ): Boolean {
+        return location.countryCode.equals("DK", ignoreCase = true)
+    }
+    override val airQualityAttribution = null
+    override val pollenAttribution = null
+    override val minutelyAttribution = null
+    override val alertAttribution = weatherAttribution
+    override val normalsAttribution = null
+
+    override fun requestSecondaryWeather(
+        context: Context, location: Location,
+        requestedFeatures: List<SecondaryWeatherSourceFeature>
+    ): Observable<SecondaryWeatherWrapper> {
+        if (!isFeatureSupportedForLocation(SecondaryWeatherSourceFeature.FEATURE_ALERT, location)) {
+            // TODO: return Observable.error(UnsupportedFeatureForLocationException())
+            return Observable.error(SecondaryWeatherException())
+        }
+
+        val id = location.parameters.getOrElse(id) { null }?.getOrElse("id") { null }
+        if (id.isNullOrEmpty()) {
+            return Observable.error(SecondaryWeatherException())
+        }
+        return mApi.getAlerts(id).map {
+            convertSecondary(it)
+        }
+    }
+
+    // Reverse geocoding
     override fun requestReverseGeocodingLocation(
         context: Context,
         location: Location
     ): Observable<List<Location>> {
-        val dmiResult = mApi.getWeather(
+        return mApi.getWeather(
             location.latitude,
             location.longitude,
-            "llj"
-        )
-
-        return dmiResult.map {
+            DMI_WEATHER_CMD
+        ).map {
             val locationList: MutableList<Location> = ArrayList()
             locationList.add(convert(location, it))
             locationList
         }
     }
 
+    // Location parameters
+    override fun needsLocationParametersRefresh(
+        location: Location,
+        coordinatesChanged: Boolean,
+        features: List<SecondaryWeatherSourceFeature>
+    ): Boolean {
+        // If we are in Denmark, we need location parameters (update it if coordinates changes,
+        // OR if we didn't have it yet)
+        return location.countryCode.equals("DK", ignoreCase = true)
+                && (coordinatesChanged || location.parameters
+                    .getOrElse(id) { null }?.getOrElse("id") { null }.isNullOrEmpty())
+    }
+
+    override fun requestLocationParameters(
+        context: Context, location: Location
+    ): Observable<Map<String, String>> {
+        return mApi.getWeather(
+            location.latitude,
+            location.longitude,
+            DMI_WEATHER_CMD
+        ).map {
+            if (it.id.isNullOrEmpty())  {
+                throw InvalidLocationException()
+            }
+            mapOf(
+                "id" to it.id
+            )
+        }
+    }
+
     companion object {
         private const val DMI_BASE_URL = "https://www.dmi.dk/"
+        private const val DMI_WEATHER_CMD = "llj"
     }
 }