+++ app/src/main/java/org/breezyweather/sources/here/json/HereWeatherPlaceReport.kt
@@ -27,6 +27,5 @@ data class HereWeatherPlaceReport(
     val observations: List<HereWeatherData>?,
     val dailyForecasts: List<HereWeatherForecasts<HereWeatherData>>?,
     val hourlyForecasts: List<HereWeatherForecasts<HereWeatherData>>?,
-    val astronomyForecasts: List<HereWeatherForecasts<HereWeatherAstronomy>>?,
-    val nwsAlerts: HereWeatherNWSAlerts?
+    val astronomyForecasts: List<HereWeatherForecasts<HereWeatherAstronomy>>?
 )